Notification tables - api calling optimization
